moving nistoar.pdr.publish.mdserv.config to nistoar.pdr.config
